drastic performance boost implemented
To clip relevant geometry features to a given tile,
a geometrical intersection needs to be performed.
By replacing the PostGIS Function 'ST_Intersect' with '&&'
in each intersection operation, query time is drastically reduced.
This is because ST_Intersect performs an exact intersection,
while '&&' only compares bounding boxes. In the case of vector tiles,
finding the intersection of bounding boxes is completely sufficient.