Auto-Reprojection of source geometries included
This is achieved by creating a temporary table
with the union of all geometries across all specified layers.
Within this process, the geoms are reprojected into the target
srid. This also should be a performance improvement over
selecting the geom union and reprojecting it on each individual tile generation query.